Move font utils to their own files
