+++ Droid/Qoden.UI.Android.csproj
@@ -48,6 +48,8 @@
   <ItemGroup>
     <Compile Include="Resources\Resource.designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="src\Graphics\CustomTypefaceSpan.cs" />
+    <Compile Include="src\Graphics\SpannableStringExtensions.cs" />
     <Compile Include="src\RemoteImageView.cs" />
     <Compile Include="src\IReverseAdapter.cs" />
     <Compile Include="src\QodenActivity.cs" />
